// components/study/types.ts
export interface StudyItemContext {
  fen?: string
  userMove?: string
  bestMove?: string
  centipawnLoss?: number
  solution?: string
  notes?: string
  moves?: string
  mainIdea?: string
  typicalPlan?: string
  dangerousIdeas?: string
  motif?: string
  theme?: string
  goal?: string
  phase?: string
}